requirements.txt
# Core Dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pyephem>=4.1.5
pytz>=2024.1

# Optional Dependencies for Enhanced Features
redis>=5.0.0

# Development Dependencies
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
